refactor Wallet type out of base
