+++ packages/core/base/src/index.ts
@@ -1,4 +1,4 @@
 export * from './adapter';
 export * from './errors';
 export * from './signer';
-export * from './wallet';
+export * from './types';
